Add Main Menu
Adds a main menu with a functioning button to start and quit the game. On my end, I set this as the main scene in Godot (the one that first plays upon start-up), and to avoid confusion I renamed main.gd and main.tscn to level.gd and level.tscn respectively. We'll need to make those same edits to the main branch for the main menu to work properly.