Non-Functional Shop UI
Adds a non-functional Shop UI with lists to select a weapon and equipment item. Currently, the shop doesn't equip anything to the player. The UI allows the player to confirm their selection before entering the game level. A similar confirm window has also been added to the main menu for when you quit the game.