Implement Functional Shop
Adds a singleton that stores the indices of the weapon/equipment items a player has selected from the shop. These indices are accessed by the player instance upon start-up to determine which weapon to preload. This has not yet been implemented for equipment items since there is no functioning equipment in the game yet.